// src/api-client/models/OtherExpensesResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Budget Manager API
 * Personal finance management API for tracking income, expenses, accounts, and budgets.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { BudgetSectionItemDto } from './BudgetSectionItemDto';
import {
    BudgetSectionItemDtoFromJSON,
    BudgetSectionItemDtoFromJSONTyped,
    BudgetSectionItemDtoToJSON,
    BudgetSectionItemDtoToJSONTyped,
} from './BudgetSectionItemDto';

/**
 * 
 * @export
 * @interface OtherExpensesResponseDto
 */
export interface OtherExpensesResponseDto {
    /**
     * 
     * @type {Array<BudgetSectionItemDto>}
     * @memberof OtherExpensesResponseDto
     */
    items?: Array<BudgetSectionItemDto> | null;
    /**
     * 
     * @type {number}
     * @memberof OtherExpensesResponseDto
     */
    total?: number;
}

/**
 * Check if a given object implements the OtherExpensesResponseDto interface.
 */
export function instanceOfOtherExpensesResponseDto(value: object): value is OtherExpensesResponseDto {
    return true;
}

export function OtherExpensesResponseDtoFromJSON(json: any): OtherExpensesResponseDto {
    return OtherExpensesResponseDtoFromJSONTyped(json, false);
}

export function OtherExpensesResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): OtherExpensesResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'items': json['items'] == null ? undefined : ((json['items'] as Array<any>).map(BudgetSectionItemDtoFromJSON)),
        'total': json['total'] == null ? undefined : json['total'],
    };
}

export function OtherExpensesResponseDtoToJSON(json: any): OtherExpensesResponseDto {
    return OtherExpensesResponseDtoToJSONTyped(json, false);
}

export function OtherExpensesResponseDtoToJSONTyped(value?: OtherExpensesResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'items': value['items'] == null ? undefined : ((value['items'] as Array<any>).map(BudgetSectionItemDtoToJSON)),
        'total': value['total'],
    };
}